feat: setup preview inline editing
(cherry picked from commit f611284285f8376c3ff2a42d449c3b0ad1e9ad83)
