+++ next/app/[locale]/(marketing)/ClientSlugHandler.tsx
@@ -18,26 +18,5 @@ export default function ClientSlugHandler({
     }
   }, [localizedSlugs, dispatch]);
 
-  const router = useRouter();
-
-  useEffect(() => {
-    const handleMessage = async (message: MessageEvent<any>) => {
-      if (
-        message.origin === process.env.NEXT_PUBLIC_API_URL &&
-        message.data.type === 'strapiUpdate'
-      ) {
-        router.refresh();
-      }
-    };
-
-    // Add the event listener
-    window.addEventListener('message', handleMessage);
-
-    // Cleanup the event listener on unmount
-    return () => {
-      window.removeEventListener('message', handleMessage);
-    };
-  }, [router]);
-
   return null; // This component only handles the state and doesn't render anything.
 }
